+++ pvnet_app/consts.py
@@ -1,0 +1,3 @@
+sat_path = "sat.zarr"
+sat_15_path = "sat_15.zarr"
+nwp_path = "nwp.zarr"
